+++ functions/src/functions/index.ts
@@ -1,9 +1,7 @@
 import * as functions from "firebase-functions";
 import * as express from "express";
 import * as cors from "cors";
-import MeetingRouter from "./meeting";
-
-import UserRouter from "./user";
+import MeetingRouters from "./meetingRouters";
 
 const corsMiddleware = cors({ origin: process.env.CORS_ORIGIN });
 
@@ -11,7 +9,6 @@ const app = express();
 
 app.use(corsMiddleware);
 
-app.use("/meetings", MeetingRouter);
-app.use("/meetings", UserRouter);
+app.use("/meetings", MeetingRouters);
 
 export const v1 = functions.https.onRequest(app);
