+++ functions/src/functions/meetingRouters/meeting.ts
@@ -1,8 +1,8 @@
 import { plainToInstance } from "class-transformer";
 import { validateOrReject } from "class-validator";
 import * as functions from "firebase-functions";
-import { CreateMeetingDto } from "../dtos/meetings";
-import { createMeeting, findMeeting } from "../services/meetings";
+import { CreateMeetingDto } from "../../dtos/meetings";
+import { createMeeting, findMeeting } from "../../services/meetings";
 import * as express from "express";
 
 const router = express.Router();
@@ -41,7 +41,9 @@ router.post(`/`, async (req, res) => {
   }
 
   // Call meetings service to create meeting
-  const { password, ...createdMeetingWithoutPassword } = await createMeeting(createMeetingDto);
+  const { password, ...createdMeetingWithoutPassword } = await createMeeting(
+    createMeetingDto
+  );
 
   // Return created meeting in proper format
   res.send(createdMeetingWithoutPassword);
